+++ extensions/tidyread---streamline-your-daily-reading/CHANGELOG.md
@@ -1,8 +1,8 @@
 # TidyRead - Streamline Your Daily Reading Changelog
 
-## [Support Quickly Generating Sources By User Interest] - 2024-01-24
+## [Onboarding Page] - 2024-01-24
 - add quickly generating sources by user interest
-- add retryDelay setting to decrease article summarizing error 
+- add retryCount and retryDelay setting to decrease article summarizing error 
 - improve speed of batch import sources
 - improve error handling and error message
 - add Delete All Sources